LLMs/evals/requirements.txt
python-dotenv
requests
beautifulsoup4
html2text
openai
youtube-transcript-api
rouge-score